finishing

+++ src/com/suman/BD/HybridMethod/HybridAlgorithmImplementation.java
@@ -2,7 +2,7 @@ package com.suman.BD.HybridMethod;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -26,7 +26,7 @@ public class HybridAlgorithmImplementation {
 		job.setMapperClass(MapperClass.class);
 		job.setReducerClass(ReducerClass.class);
 		job.setMapOutputKeyClass(Pair.class);
-		job.setMapOutputValueClass(IntWritable.class);
+		job.setMapOutputValueClass(DoubleWritable.class);
 		job.setOutputKeyClass(Text.class);
 		job.setOutputValueClass(CustomMapWritable.class);
 		System.exit(job.waitForCompletion(true) ? 0 : 1);
